requirements.txt
# ── Core services ──────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# ── LLM backends ───────────────────────────────────────────────────────────
anthropic>=0.25.0          # Anthropic API (Phase 1 dev)
# ollama is accessed via its OpenAI-compatible HTTP API — no package needed

# ── Vector DB + embeddings ─────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=2.7.0

# ── Token counting ─────────────────────────────────────────────────────────
tiktoken>=0.7.0

# ── Discord ────────────────────────────────────────────────────────────────
discord.py>=2.3.0

# ── Config ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
